fixes issue #23
